+++ src/Commands/ServeCommand.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Anteris\Selenium\Commands;
+namespace Anteris\Selenium\Server\Commands;
 
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Exception\InvalidOptionException;