+++ src/Exceptions/InstallException.php
@@ -1,0 +1,6 @@
+<?php
+
+namespace Anteris\Selenium\Server\Exceptions;
+
+class InstallException extends \Exception
+{}