+++ .github/CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased] - 2020-09-16
+
+### Added
+- Automatic download of Gecko and Chrome binaries upon install so that they don't slow down install time
+
+### Changed
+- Namespace from `Anteris\Selenium` to `Anteris\Selenium\Server`
+
+### Removed
+- Selenium binaries from `/bin`
+
 ## [v0.1.0] - 2020-09-15
 
 ### Added